Move dump-heads into its own tool
