+++ cmd/promtool/main.go
@@ -24,7 +24,6 @@ import (
 
 	"github.com/prometheus/prometheus/config"
 	"github.com/prometheus/prometheus/promql"
-	"github.com/prometheus/prometheus/storage/local"
 	"github.com/prometheus/prometheus/util/cli"
 	"github.com/prometheus/prometheus/version"
 )
@@ -185,19 +184,6 @@ func checkRules(t cli.Term, filename string) (int, error) {
 	return len(rules), nil
 }
 
-// DumpHeadsCmd dumps metadata of a heads.db file.
-func DumpHeadsCmd(t cli.Term, args ...string) int {
-	if len(args) != 1 {
-		t.Infof("usage: promtool dump-heads <file>")
-		return 2
-	}
-	if err := local.DumpHeads(args[0], t.Out()); err != nil {
-		t.Errorf("  FAILED: %s", err)
-		return 1
-	}
-	return 0
-}
-
 var versionInfoTmpl = `
 prometheus, version {{.version}} (branch: {{.branch}}, revision: {{.revision}})
   build user:       {{.buildUser}}
@@ -228,11 +214,6 @@ func main() {
 	app.Register("check-rules", &cli.Command{
 		Desc: "validate rule files for correctness",
 		Run:  CheckRulesCmd,
-	})
-
-	app.Register("dump-heads", &cli.Command{
-		Desc: "dump metadata of a heads.db checkpoint file",
-		Run:  DumpHeadsCmd,
 	})
 
 	app.Register("version", &cli.Command{
